refactor: extract QuoteErrorsButton pure component
